refactor src to have features and shared sub directories and place appropriate files accordingly, refactoring where appropriate
